moved the espresso tests
