+++ frontend/app/src/androidTest/java/com/example/shopeer/BrowseManagePeersTest.java
@@ -1,4 +1,4 @@
-package espressotests;
+package com.example.shopeer;
 
 import com.android.volley.toolbox.JsonArrayRequest;
 
@@ -39,9 +39,6 @@ import com.android.volley.VolleyError;
 import com.android.volley.toolbox.JsonObjectRequest;
 import com.android.volley.toolbox.StringRequest;
 import com.android.volley.toolbox.Volley;
-import com.example.shopeer.ChatActivity;
-import com.example.shopeer.MainActivity;
-import com.example.shopeer.R;
 
 import org.json.JSONArray;
 import org.json.JSONException;
