+++ frontend/app/src/androidTest/java/com/example/shopeer/RecyclerViewUtil.java
@@ -1,4 +1,4 @@
-package espressotests;
+package com.example.shopeer;
 
 import android.content.res.Resources;
 import android.view.View;
